add legacy redirect and stub new sign page based off proposals
