+++ packages/db/src/access/transactions/index.ts
@@ -1,2 +1,4 @@
+export { getProposedTransactions } from './getProposedTransactions'
 export { getTransactions } from './getTransactions'
+export { proposeTransaction } from './proposeTransaction'
 export { saveTransaction } from './saveTransaction'
